Support aspect ration for each map
We do not want to *require* webmasters to use CSS to set the height of
the maps, nor do we want to hard-code this.  And actually, the height
is not a good fit for responsive design, or generally for different
placement (e.g. in the main contents of a page, or a sidebar in the
template).

Therefore, we require users to specify the aspect ratio for each map.