Support multiple maps on a page
This can make sense in some cases; it can even make sense to present
the same map multiple times (e.g. in a sidebar in the template, but
also enlarged on a page).

Therefore we no longer use a hard-coded ID, but rather walk over all
`figure.maps_map`s and initialize the maps.