Require PHP ≥ 8.0.0
Prior to this PHP version, float to string conversion is locale-
dependent what is total non-sense, and it's painful to work-around.
Therefore we state that we require PHP 8, although the plugin should
work fine with 7.4 and an English locale (or any other locale where
the decimal separator is a period).