Use form serialization instead of manually constructing markers
This is more robust regarding adding further fields (what may happen at
some point in the future), and also looks cleaner.  Since `FormData`
isn't as flexible as jQuery's `serialize()`, we append a clone of the
table to a newly created form, so we can do the serialization.  To
avoid submitting the additional useless fields (which may actually
exceed PHP's `max_input_vars`), we empty their names before we let the
client submit the form.