Convert closure based widget to object
We're barely used to work and maintain code with nested functions, but
are used to work and maintain object oriented code, so this is mostly a
style preference.  Particularly, we like to be able to use getters
instead of vars or even functions, and actually like that we are
pressed to extract small methods to cleanly solve the `this` bindings.